Create screens and navigation tree with props and components
We need screens to navigate between logging in and displaying the API response,
but to do that we installed the React Navigation and React Native Screens libraries.
This commit adds the Login screen, Button component to be reused for the simple UI,
and the navigation file to keep the necessary typings.