Fetch the profile from GitHub API and display it in Profile screen
To display the results of fetching the API we introduced the new screen:
Profile,
which is filled with everything that is received as GitHubAccount type of object.
We only navigate to that screen if we have all states settled:
authCode, accessToken, and profile.
The latter - profile - is fetched in an effect that launches the handleAPIFetch
callback as soon as both authCode and accessToken are received.